moved tests around
