+++ RUFAS/routines/field/manager/field_manager.py
@@ -49,7 +49,7 @@ class FieldManager:
 
         """
         for field in self.fields:
-            month = FieldManager._date_conversion_month(time)
+            month = CurrentWeather.date_conversion_month(time)
             current_weather = FieldManager._create_current_weather(weather=weather, time=time, month=month)
             field.manage_field(time, current_weather=current_weather)
         self.output_gatherer.send_daily_variables()
@@ -62,61 +62,6 @@ class FieldManager:
         self.output_gatherer.send_annual_variables()
         for field in self.fields:
             field.perform_annual_reset()
-
-    @staticmethod
-    def _date_conversion_month(time) -> int:
-        """
-        Converts the day number into the corresponding month of the year.
-
-        Parameters
-        ----------
-        time: Time
-            Object containing the current year and day of the simulation.
-
-        Returns
-        -------
-        int
-            The corresponding month of the year.
-
-        """
-        days = [31, 59, 90, 120, 151, 181, 212, 243, 273, 304, 334, 365]
-        leap_days = [31, 60, 91, 121, 152, 182, 213, 244, 274, 305, 335, 366]
-        prev_month = 0
-        if Schedule.is_leap_year(time.calendar_year):
-            for day in leap_days:
-                if prev_month < time.day <= day:
-                    return leap_days.index(day) + 1
-                else:
-                    prev_month = day
-        else:
-            for day in days:
-                if prev_month < time.day <= day:
-                    return days.index(day) + 1
-                else:
-                    prev_month = day
-
-    @staticmethod
-    def _date_conversion_day(time) -> int:
-        """
-        Converts the day number into the corresponding day of the month.
-        Parameters
-        ----------
-        time:
-            Object containing the current year and day of the simulation.
-
-        Returns
-        -------
-        int
-        Corresponding day of the month.
-
-        """
-        days = [31, 59, 90, 120, 151, 181, 212, 243, 273, 304, 334, 365]
-        leap_days = [31, 60, 91, 121, 152, 182, 213, 244, 274, 305, 335, 366]
-
-        if Schedule.is_leap_year(time.calendar_year):
-            return time.day - leap_days[FieldManager._date_conversion_month(time) - 2]
-        else:
-            return time.day - days[FieldManager._date_conversion_month(time) - 2]
 
     @staticmethod
     def _create_current_weather(weather, time, month: int) -> CurrentWeather:
