+++ tests/classes_tests/test_weather.py
